Change base url and remove useless method
