+++ docs/enums/ErrorCodes.html
@@ -91,7 +91,7 @@
 					<div class="tsd-signature tsd-kind-icon">API_<wbr>KEY_<wbr>FAILED<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;Could not find api key in request.&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/util/constants.ts#L6">src/util/constants.ts:6</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/util/constants.ts#L6">src/util/constants.ts:6</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -101,7 +101,7 @@
 					<div class="tsd-signature tsd-kind-icon">CLIENT_<wbr>VERSION_<wbr>FAILED<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;Could not find client version in request.&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/util/constants.ts#L7">src/util/constants.ts:7</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/util/constants.ts#L7">src/util/constants.ts:7</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -111,7 +111,7 @@
 					<div class="tsd-signature tsd-kind-icon">NO_<wbr>AVAILABLE_<wbr>FORMAT<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;Could not find a suitable format for this download.&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/util/constants.ts#L2">src/util/constants.ts:2</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/util/constants.ts#L2">src/util/constants.ts:2</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -121,7 +121,7 @@
 					<div class="tsd-signature tsd-kind-icon">PLAYLIST_<wbr>LOAD_<wbr>FAILED<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;Failed to load desired playlist.&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/util/constants.ts#L4">src/util/constants.ts:4</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/util/constants.ts#L4">src/util/constants.ts:4</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -131,7 +131,7 @@
 					<div class="tsd-signature tsd-kind-icon">SEARCH_<wbr>FAILED<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;Failed to search for videos.&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/util/constants.ts#L5">src/util/constants.ts:5</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/util/constants.ts#L5">src/util/constants.ts:5</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -141,7 +141,7 @@
 					<div class="tsd-signature tsd-kind-icon">UNKNOWN_<wbr>TOKEN<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol"> = &quot;An invalid token was found.&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/util/constants.ts#L3">src/util/constants.ts:3</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/util/constants.ts#L3">src/util/constants.ts:3</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -247,9 +247,6 @@
 					<li class=" tsd-kind-interface">
 						<a href="../interfaces/YoutubeVideoFormat.html" class="tsd-kind-icon">Youtube<wbr>Video<wbr>Format</a>
 					</li>
-					<li class=" tsd-kind-property">
-						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
-					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#DEFAULT_CONTEXT" class="tsd-kind-icon">DEFAULT_<wbr>CONTEXT</a>
 					</li>
@@ -270,6 +267,9 @@
 					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#cachedTokens" class="tsd-kind-icon">cached<wbr>Tokens</a>
+					</li>
+					<li class=" tsd-kind-variable">
+						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="../modules.html#download" class="tsd-kind-icon">download</a>
