+++ docs/interfaces/DownloadOptions.html
@@ -98,7 +98,7 @@
 					<div class="tsd-signature tsd-kind-icon">begin<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeVideo.ts#L86">src/structures/YoutubeVideo.ts:86</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeVideo.ts#L86">src/structures/YoutubeVideo.ts:86</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -108,7 +108,7 @@
 					<div class="tsd-signature tsd-kind-icon">chunk<wbr>Mode<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{ </span>chunkSize<span class="tsd-signature-symbol">?: </span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> }</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeVideo.ts#L81">src/structures/YoutubeVideo.ts:81</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeVideo.ts#L81">src/structures/YoutubeVideo.ts:81</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-type-declaration">
@@ -126,7 +126,7 @@
 					<div class="tsd-signature tsd-kind-icon">high<wbr>Water<wbr>Mark<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeVideo.ts#L84">src/structures/YoutubeVideo.ts:84</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeVideo.ts#L84">src/structures/YoutubeVideo.ts:84</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -136,7 +136,7 @@
 					<div class="tsd-signature tsd-kind-icon">pipe<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeVideo.ts#L87">src/structures/YoutubeVideo.ts:87</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeVideo.ts#L87">src/structures/YoutubeVideo.ts:87</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -146,7 +146,7 @@
 					<div class="tsd-signature tsd-kind-icon">resource<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">PassThrough</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeVideo.ts#L85">src/structures/YoutubeVideo.ts:85</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeVideo.ts#L85">src/structures/YoutubeVideo.ts:85</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -249,9 +249,6 @@
 					<li class=" tsd-kind-interface">
 						<a href="YoutubeVideoFormat.html" class="tsd-kind-icon">Youtube<wbr>Video<wbr>Format</a>
 					</li>
-					<li class=" tsd-kind-property">
-						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
-					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#DEFAULT_CONTEXT" class="tsd-kind-icon">DEFAULT_<wbr>CONTEXT</a>
 					</li>
@@ -272,6 +269,9 @@
 					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#cachedTokens" class="tsd-kind-icon">cached<wbr>Tokens</a>
+					</li>
+					<li class=" tsd-kind-variable">
+						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="../modules.html#download" class="tsd-kind-icon">download</a>
