+++ docs/interfaces/PlaylistVideo.html
@@ -102,7 +102,7 @@
 					<div class="tsd-signature tsd-kind-icon">duration<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L18">src/structures/Playlist.ts:18</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L18">src/structures/Playlist.ts:18</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -112,7 +112,7 @@
 					<div class="tsd-signature tsd-kind-icon">formatted<wbr>Duration<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L19">src/structures/Playlist.ts:19</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L19">src/structures/Playlist.ts:19</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -122,7 +122,7 @@
 					<div class="tsd-signature tsd-kind-icon">formatted<wbr>Readable<wbr>Duration<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L20">src/structures/Playlist.ts:20</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L20">src/structures/Playlist.ts:20</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -132,7 +132,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L9">src/structures/Playlist.ts:9</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L9">src/structures/Playlist.ts:9</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -142,7 +142,7 @@
 					<div class="tsd-signature tsd-kind-icon">index<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L17">src/structures/Playlist.ts:17</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L17">src/structures/Playlist.ts:17</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -152,7 +152,7 @@
 					<div class="tsd-signature tsd-kind-icon">is<wbr>Playable<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L21">src/structures/Playlist.ts:21</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L21">src/structures/Playlist.ts:21</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -162,7 +162,7 @@
 					<div class="tsd-signature tsd-kind-icon">thumbnails<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{ </span>height<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol">; </span>url<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>width<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> }</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L11">src/structures/Playlist.ts:11</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L11">src/structures/Playlist.ts:11</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -172,7 +172,7 @@
 					<div class="tsd-signature tsd-kind-icon">title<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L16">src/structures/Playlist.ts:16</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L16">src/structures/Playlist.ts:16</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -182,7 +182,7 @@
 					<div class="tsd-signature tsd-kind-icon">url<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/Playlist.ts#L10">src/structures/Playlist.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/Playlist.ts#L10">src/structures/Playlist.ts:10</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -297,9 +297,6 @@
 					<li class=" tsd-kind-interface">
 						<a href="YoutubeVideoFormat.html" class="tsd-kind-icon">Youtube<wbr>Video<wbr>Format</a>
 					</li>
-					<li class=" tsd-kind-property">
-						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
-					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#DEFAULT_CONTEXT" class="tsd-kind-icon">DEFAULT_<wbr>CONTEXT</a>
 					</li>
@@ -320,6 +317,9 @@
 					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#cachedTokens" class="tsd-kind-icon">cached<wbr>Tokens</a>
+					</li>
+					<li class=" tsd-kind-variable">
+						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="../modules.html#download" class="tsd-kind-icon">download</a>
