+++ docs/interfaces/YoutubeSearchChannelInfo.html
@@ -99,7 +99,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L49">src/structures/YoutubeSearchResults.ts:49</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L49">src/structures/YoutubeSearchResults.ts:49</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -109,7 +109,7 @@
 					<div class="tsd-signature tsd-kind-icon">subscriber<wbr>Count<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L58">src/structures/YoutubeSearchResults.ts:58</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L58">src/structures/YoutubeSearchResults.ts:58</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -119,7 +119,7 @@
 					<div class="tsd-signature tsd-kind-icon">thumbnails<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{ </span>height<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>url<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>width<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> }</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L50">src/structures/YoutubeSearchResults.ts:50</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L50">src/structures/YoutubeSearchResults.ts:50</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -129,7 +129,7 @@
 					<div class="tsd-signature tsd-kind-icon">title<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L56">src/structures/YoutubeSearchResults.ts:56</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L56">src/structures/YoutubeSearchResults.ts:56</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -139,7 +139,7 @@
 					<div class="tsd-signature tsd-kind-icon">url<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L55">src/structures/YoutubeSearchResults.ts:55</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L55">src/structures/YoutubeSearchResults.ts:55</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -149,7 +149,7 @@
 					<div class="tsd-signature tsd-kind-icon">verified<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L57">src/structures/YoutubeSearchResults.ts:57</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L57">src/structures/YoutubeSearchResults.ts:57</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -255,9 +255,6 @@
 					<li class=" tsd-kind-interface">
 						<a href="YoutubeVideoFormat.html" class="tsd-kind-icon">Youtube<wbr>Video<wbr>Format</a>
 					</li>
-					<li class=" tsd-kind-property">
-						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
-					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#DEFAULT_CONTEXT" class="tsd-kind-icon">DEFAULT_<wbr>CONTEXT</a>
 					</li>
@@ -278,6 +275,9 @@
 					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#cachedTokens" class="tsd-kind-icon">cached<wbr>Tokens</a>
+					</li>
+					<li class=" tsd-kind-variable">
+						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="../modules.html#download" class="tsd-kind-icon">download</a>
