+++ docs/interfaces/YoutubeSearchListInfo.html
@@ -99,7 +99,7 @@
 					<div class="tsd-signature tsd-kind-icon">channel<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{ </span>id<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>name<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>url<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> }</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L40">src/structures/YoutubeSearchResults.ts:40</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L40">src/structures/YoutubeSearchResults.ts:40</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-type-declaration">
@@ -123,7 +123,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L32">src/structures/YoutubeSearchResults.ts:32</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L32">src/structures/YoutubeSearchResults.ts:32</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -133,7 +133,7 @@
 					<div class="tsd-signature tsd-kind-icon">thumbnails<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{ </span>height<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>url<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">; </span>width<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol"> }</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L33">src/structures/YoutubeSearchResults.ts:33</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L33">src/structures/YoutubeSearchResults.ts:33</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -143,7 +143,7 @@
 					<div class="tsd-signature tsd-kind-icon">title<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L39">src/structures/YoutubeSearchResults.ts:39</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L39">src/structures/YoutubeSearchResults.ts:39</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -153,7 +153,7 @@
 					<div class="tsd-signature tsd-kind-icon">url<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L38">src/structures/YoutubeSearchResults.ts:38</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L38">src/structures/YoutubeSearchResults.ts:38</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -163,7 +163,7 @@
 					<div class="tsd-signature tsd-kind-icon">video<wbr>Count<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/0a88e52/src/structures/YoutubeSearchResults.ts#L45">src/structures/YoutubeSearchResults.ts:45</a></li>
+							<li>Defined in <a href="https://github.com/cjh980402/youtube-scrapper/blob/be8f9dc/src/structures/YoutubeSearchResults.ts#L45">src/structures/YoutubeSearchResults.ts:45</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -269,9 +269,6 @@
 					<li class=" tsd-kind-interface">
 						<a href="YoutubeVideoFormat.html" class="tsd-kind-icon">Youtube<wbr>Video<wbr>Format</a>
 					</li>
-					<li class=" tsd-kind-property">
-						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
-					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#DEFAULT_CONTEXT" class="tsd-kind-icon">DEFAULT_<wbr>CONTEXT</a>
 					</li>
@@ -292,6 +289,9 @@
 					</li>
 					<li class=" tsd-kind-variable">
 						<a href="../modules.html#cachedTokens" class="tsd-kind-icon">cached<wbr>Tokens</a>
+					</li>
+					<li class=" tsd-kind-variable">
+						<a href="../modules.html#formats" class="tsd-kind-icon">formats</a>
 					</li>
 					<li class=" tsd-kind-function">
 						<a href="../modules.html#download" class="tsd-kind-icon">download</a>
