+++ src/util/formats.ts
@@ -1,15 +1,7 @@
 /**
  * http://en.wikipedia.org/wiki/YouTube#Quality_and_formats
  */
-
-export interface RawVideoFormat {
-    mimeType: string;
-    qualityLabel: string | null;
-    bitrate: number | null;
-    audioBitrate?: number | null;
-}
-
-export default {
+export const formats = {
     5: {
         mimeType: 'video/flv; codecs="Sorenson H.283, mp3"',
         qualityLabel: '240p',
@@ -527,4 +519,4 @@ export default {
         bitrate: 12000000,
         audioBitrate: null
     }
-} as Record<number, RawVideoFormat>;
+};
