+++ src/util/index.ts
@@ -1,7 +1,6 @@
-export { default as formats } from './formats';
-
 export * from './cache';
 export * from './constants';
+export * from './formats';
 export * from './noop';
 export * from './Regexes';
 export * from './Util';
